Ensuring that when init is run, we are ready to apply other tf commands
